refactor(cli): remove skill-token module and integrate functionality into pi-tui-pickers
This commit deletes the skill-token module and moves the skill invocation prefix logic into pi-tui-pickers. The skill invocation prefix function is now defined directly in pi-tui-pickers, streamlining the codebase. Additionally, the import statement for SKILL_INVOCATION_TOKEN_SOURCE has been updated in skill-highlight-editor to reflect the new structure.
